+++ my-app/Backend/routes/commentRoutes.js
@@ -7,6 +7,7 @@ const commentController = require('../controllers/commentController');
 router.get("/", commentController.getComments);       
 router.post("/", commentController.createComment);    
 router.delete("/:id", commentController.deleteComment); 
+router.put("/:id", commentController.updateComment);
 
 
 module.exports = router;
